PO-2254 report instance ID update

+++ src/main/java/uk/gov/hmcts/opal/controllers/ReportInstancesApiController.java
@@ -51,7 +51,7 @@ public class ReportInstancesApiController implements ReportInstancesApi {
 
 
     @Override
-    public ResponseEntity<ReportInstanceReports> getReportInstance(Integer id) {
+    public ResponseEntity<ReportInstanceReports> getReportInstance(Long id) {
         return buildResponse(genericReportService.getReportInstance(id));
     }
 }
